Add new file for test HelloWorld component
